rag_service/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-chroma
langchain-ollama
langchain-google-genai
chromadb
polygon-api-client
python-dotenv